apps/desktop/01_ACTIVE_SYSTEMS/neuropetrix_working_backup/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-docx==1.1.2
rapidfuzz==3.9.6
